refactor: moved FloorSidebar into it's own file (later planning on making a AreaSidebar.tsx)
